// simni-next/src/types/settings.ts
export interface SchoolSettings {
  nama_aplikasi: string;
  tahun_pelajaran: string;
  nama_yayasan: string;
  nama_sekolah: string;
  status_akreditasi: string;
  kota: string;
  nomor_izin: string;
  kepala_sekolah?: string;
  nuks_kepala_sekolah?: string;
  wali_kelas?: string;
  nuptk_wali_kelas?: string;
}